# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Application Programming Interface (API) for Pasiap Apps",
        "description": "L5 Swagger OpenAPI documentation for Pasiap Apps",
        "contact": {
            "email": "burhanburdev@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/pasiap/public/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/faq": {
            "get": {
                "tags": [
                    "Faq"
                ],
                "description": "Get all data FAQ",
                "operationId": "getAllFaq",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Faq"
                ],
                "description": "Add faq",
                "operationId": "storeFaq",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "question": {
                                        "type": "string"
                                    },
                                    "answer": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/faq/{id}": {
            "delete": {
                "tags": [
                    "Faq"
                ],
                "description": "Delete faq by ID",
                "operationId": "deleteFaq",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the data to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}